backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-prometheus
django-stubs
channels
channels-redis
daphne
psycopg[binary]
redis
gmqtt
pydantic
python-dotenv
pyjwt
pytest
pytest-django
pytest-asyncio
pytest-cov
pytest-mock
ruff
mypy
